% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{cd_info}
\alias{cd_info}
\title{District 2018 Descriptions and Information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 435 rows and 9 columns.
}
\source{
The Daily Kos Elections naming guide to the nation's congressional districts.
\url{https://bit.ly/2XsFI5W}

Daily Kos, "2008, 2012, & 2016 results for districts used in 2018."
\url{https://bit.ly/3bXtAPB}
}
\usage{
cd_info
}
\description{
District 2018 Descriptions and Information
}
\seealso{
\url{https://www.shirokuriwaki.com/ccesMRPprep/reference/cd_info_2018.html}
}
\author{
Daily Kos Elections
}
\keyword{datasets}
