% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp.R
\docType{data}
\name{cd_shp}
\alias{cd_shp}
\title{Congressional District Cartogram Shapefile}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 435 rows and 6 columns.
}
\usage{
cd_shp
}
\description{
Congressional District Cartogram Shapefile
}
\examples{
library(ggplot2)
library(ggthemes)

ggplot(cd_shp) + geom_sf() + theme_map()

library(dplyr)
cd_black <- left_join(cd_shp, race_by_cd, by = "cd") \%>\%
  filter(race == "Black")

ggplot(cd_black) +
  geom_sf(aes(fill = frac)) +
  scale_fill_viridis_b("Percent Black") +
  theme_map() +
  labs(caption = "Map Source: http://dkel.ec/map")



}
\author{
Daniel Donner \url{http://dkel.ec/map}
}
\keyword{datasets}
