% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp.R
\docType{data}
\name{st_shp}
\alias{st_shp}
\title{State Cartogram Shapefile}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 50 rows and 3 columns.
}
\usage{
st_shp
}
\description{
State Cartogram Shapefile
}
\examples{
library(ggplot2)
library(ggthemes)

ggplot(st_shp) + geom_sf() + theme_map()

}
\author{
Daniel Donner \url{http://dkel.ec/map}
}
\keyword{datasets}
