% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{race_by_cd}
\alias{race_by_cd}
\title{Counts and fractions of race by congressional district}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2195 rows and 5 columns.
}
\usage{
race_by_cd
}
\description{
Counts and fractions of race by congressional district
}
\author{
5-Year American Community Survey, 2014-2018
}
\keyword{datasets}
